+++ bin/main.dart
@@ -1,4 +1,5 @@
 import 'dart:io';
+
 import 'package:args/command_runner.dart';
 import 'package:flutter_translation_sheet/flutter_translation_sheet.dart';
 
@@ -27,15 +28,22 @@ Future<void> startFetch() async {
   trace('Creating local canonical json');
   var map = buildLocalYamlMap();
   var canoMap = buildCanoMap(map);
+  // trace("Map is: ", canoMap);
+  // exit(0);
+  buildVarsInMap(canoMap);
+  // var _tmp = {'en':canoMap};
+  // putVarsInMap(_tmp);
   trace('Fetching data from Google sheets...');
   final localesMap = await sheet.getData();
   localesMap[config.masterLocale] = canoMap;
   putVarsInMap(localesMap);
-
   if (config.validTKeyFile) {
     createTKeyFileFromMap(map, save: true, includeToString: true);
   }
   createLocalesFiles(localesMap);
+  if (config.intlEnabled) {
+    buildArb(localesMap);
+  }
   formatDartFiles();
   exit(1);
 }
@@ -62,6 +70,9 @@ Future<void> build() async {
     createTKeyFileFromMap(map, save: true, includeToString: true);
   }
   createLocalesFiles(localesMap);
+  if (config.intlEnabled) {
+    buildArb(localesMap);
+  }
   formatDartFiles();
   exit(1);
 }