+++ lib/src/io/io.dart
@@ -4,5 +4,6 @@ export 'commands.dart';
 export 'dart_gen.dart';
 export 'env_parser.dart';
 export 'extract_strings.dart';
+export 'intl_dart_gen.dart';
 export 'sheet_parser.dart';
 export 'yaml_parser.dart';